refactor: Abstract the processtracer and auditor modules to collect events for behavior modeling and violation auditing features
