+++ internal/agent/agent.go
@@ -38,16 +38,16 @@ import (
 	// listerv1 "k8s.io/client-go/listers/core/v1"
 	varmor "github.com/bytedance/vArmor/apis/varmor/v1beta1"
 	varmorbehavior "github.com/bytedance/vArmor/internal/behavior"
-	varmortracer "github.com/bytedance/vArmor/internal/behavior/tracer"
 	varmorconfig "github.com/bytedance/vArmor/internal/config"
 	varmortypes "github.com/bytedance/vArmor/internal/types"
 	varmorutils "github.com/bytedance/vArmor/internal/utils"
-	varmoraudit "github.com/bytedance/vArmor/pkg/audit"
+	varmorauditor "github.com/bytedance/vArmor/pkg/auditor"
 	varmorinterface "github.com/bytedance/vArmor/pkg/client/clientset/versioned/typed/varmor/v1beta1"
 	varmorinformer "github.com/bytedance/vArmor/pkg/client/informers/externalversions/varmor/v1beta1"
 	varmorlister "github.com/bytedance/vArmor/pkg/client/listers/varmor/v1beta1"
 	varmorapparmor "github.com/bytedance/vArmor/pkg/lsm/apparmor"
 	varmorbpfenforcer "github.com/bytedance/vArmor/pkg/lsm/bpfenforcer"
+	varmorptracer "github.com/bytedance/vArmor/pkg/processtracer"
 	varmorruntime "github.com/bytedance/vArmor/pkg/runtime"
 	varmorseccomp "github.com/bytedance/vArmor/pkg/seccomp"
 )
@@ -69,7 +69,7 @@ type Agent struct {
 	appArmorProfileDir       string
 	seccompProfileDir        string
 	bpfEnforcer              *varmorbpfenforcer.BpfEnforcer
-	auditor                  *varmoraudit.Auditor
+	auditor                  *varmorauditor.Auditor
 	monitor                  *varmorruntime.RuntimeMonitor
 	waitExistingApSync       sync.WaitGroup
 	existingApCount          int
@@ -78,7 +78,7 @@ type Agent struct {
 	enableBpfEnforcer        bool
 	unloadAllAaProfiles      bool
 	removeAllSeccompProfiles bool
-	tracer                   *varmortracer.Tracer
+	ptracer                  *varmorptracer.ProcessTracer
 	modellers                map[string]*varmorbehavior.BehaviorModeller
 	nodeName                 string
 	debug                    bool
@@ -237,8 +237,8 @@ func NewAgent(
 		}
 	}
 
-	// Create an auditor to audit violation behaviors for AppArmor & BPF enforcers
-	agent.auditor, err = varmoraudit.NewAuditor(agent.nodeName, agent.appArmorSupported, agent.bpfLsmSupported, log.WithName("AUDIT-VIOLATIONS"))
+	// Create an auditor to audit violation and behavior events for AppArmor, Seccomp and BPF enforcers
+	agent.auditor, err = varmorauditor.NewAuditor(agent.nodeName, agent.appArmorSupported, agent.bpfLsmSupported, log.WithName("AUDIT-VIOLATIONS"))
 	if err != nil {
 		return nil, err
 	}
@@ -246,11 +246,13 @@ func NewAgent(
 	// Subscribe auditor to the monitor
 	agent.monitor.AddTaskNotifyChs("AUDIT-VIOLATIONS", &agent.auditor.TaskStartCh, &agent.auditor.TaskDeleteCh, &agent.auditor.TaskDeleteSyncCh)
 
-	// [Experimental feature] Initialize the tracer for BehaviorModeling mode.
-	// It only works with AppArmor LSM and Seccomp for now.
+	// [Experimental feature]
+	//     Initialize the process tracer for BehaviorModeling mode.
+	//     It only works with AppArmor and Seccomp enforcers for now.
+	// TODO: Support BPF enforcer
 	if agent.enableBehaviorModeling {
-		log.Info("initialize the tracer for BehaviorModeling mode")
-		agent.tracer, err = varmortracer.NewTracer(log.WithName("TRACER"))
+		log.Info("initialize the process tracer for BehaviorModeling mode")
+		agent.ptracer, err = varmorptracer.NewProcessTracer(log.WithName("TRACER"))
 		if err != nil {
 			return nil, err
 		}
@@ -398,9 +400,10 @@ func (agent *Agent) handleCreateOrUpdateArmorProfile(ap *varmor.ArmorProfile, ke
 				modeller.PreprocessAndSendBehaviorData()
 			}
 		} else {
-			// Create a new modeller and start modeling.
+			// Create a new modeller and start modeling for the ArmorProfile object.
 			modeller := varmorbehavior.NewBehaviorModeller(
-				agent.tracer,
+				agent.auditor,
+				agent.ptracer,
 				agent.monitor,
 				agent.nodeName,
 				ap.Namespace,
@@ -664,7 +667,7 @@ func (agent *Agent) CleanUp() {
 	agent.auditor.Close()
 
 	if agent.enableBehaviorModeling {
-		agent.tracer.Close()
+		agent.ptracer.Close()
 	}
 
 	if agent.appArmorSupported && agent.unloadAllAaProfiles {
