+++ pkg/auditor/types.go
